Rename new_resource.py to resource_v2.py
