+++ checkbox-ng/plainbox/impl/resource_v2.py
